Refactor Mac

+++ Socket.cpp
@@ -2,8 +2,7 @@
 //  Socket.cpp
 //  VDJartnet
 //
-//  Created by Jonathan Tanner on 09/02/2017.
-//  Copyright © 2017 Jonathan Tanner. All rights reserved.
+//  Copyright © 2017-18 Jonathan Tanner. All rights reserved.
 //
 //This file is part of VDJartnet.
 //
@@ -29,6 +28,8 @@
 //If you modify this Program, or any covered work, by linking or
 //combining it with the Visual C++ Runtime, the licensors of this Program grant you
 //additional permission to convey the resulting work.
+//Corresponding Source for a non-source form of such a combination shall not
+//include the source code for the parts of the Visual C++ Runtime used as well as that of the covered work.
 
 #include "Socket.hpp"
 
@@ -107,7 +108,7 @@ void Socket::send(std::string hostS, unsigned short port, const void* data, int 
     address.sin_addr.s_addr = host;
     address.sin_port = htons(port);
 
-    ssize_t sent_bytes = sendto(handle, (const char *) data, size, 0, (const struct sockaddr *) &address, sizeof(struct sockaddr_in));
+    ssize_t sent_bytes = sendto(handle, (const char *) data, (size_t)size, 0, (const struct sockaddr *) &address, sizeof(struct sockaddr_in));
     if (sent_bytes != size) {
         throw std::runtime_error("Failed to send data");
     }
